+++ Changes.md
@@ -7,6 +7,7 @@ Improvements
 - SceneInspector :
   - Added "Isolate Differences" option for comparison modes. This filters out all properties which have the same value in the A and B columns.
   - Removed redundant scene inspections when not in comparison mode.
+- Arnold : Added support for M44f `header:*` output metadata.
 
 Fixes
 -----